+++ Backend/src/main/java/edu/cit/hapongo/repository/LessonContentRepository.java
@@ -4,6 +4,6 @@ import edu.cit.hapongo.model.LessonContent;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface LessonContentRepository extends JpaRepository<LessonContent, Integer> {
-    List<LessonContent> findByLesson_LessonId(int lessonId);
+public interface LessonContentRepository extends JpaRepository<LessonContent, Long> {
+    List<LessonContent> findByLesson_LessonId(long lessonId);
 }
